Rename DialogProvider to match others
