+++ src/contexts/index.js
@@ -1,6 +1,6 @@
 import { Auth0Context, Auth0Provider } from './AuthProvider';
 import { ThemeContext, ThemeProvider } from './ThemeProvider';
-import { DialogContext, DialogProvider } from './DialogContext';
+import { DialogContext, DialogProvider } from './DialogProvider';
 
 export {
   Auth0Context as AuthContext,
