+++ src/dialogs/ProjectSettingsDialog.jsx
@@ -6,7 +6,6 @@ import {
   Grid,
   useMediaQuery,
   useTheme,
-  makeStyles,
   Button,
   Dialog,
   DialogActions,
@@ -30,41 +29,7 @@ import {
   AvailablePage,
   CurrentPage
 } from '../pages/Project/components';
-
-const useStyles = makeStyles((theme) => ({
-  dialogContent: {
-    paddingTop: theme.spacing(2),
-    minWidth: 780,
-    minHeight: 560,
-    maxHeight: 560
-  },
-  headerPaper: {
-    borderTop: 'none',
-    borderLeft: 'none',
-    borderRight: 'none'
-  },
-  availableModules: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'flex-start',
-    justifyContent: 'flex-start'
-  },
-  card: {
-    width: 175,
-    borderRadius: 4
-  },
-  switch: {
-    display: 'flex',
-    justifyContent: 'flex-end'
-  }
-}));
-
-const getListStyle = (isDraggingOver) => ({
-  minHeight: '500px',
-  maxHeight: '500px',
-  overflowY: 'auto',
-  background: isDraggingOver ? 'lightblue' : ''
-});
+import { useStyles, getListStyle } from './ProjectSettingsDialogStyles';
 
 export default function ProjectSettingsDialog() {
   const classes = useStyles();
